small code reorganization
